+++ bathbot/src/commands/osu/nochoke.rs
@@ -440,7 +440,7 @@ async fn process_scores(
             map,
             max_pp,
             stars: attrs.stars() as f32,
-            max_combo: attrs.max_combo() as u32,
+            max_combo: attrs.max_combo(),
         };
 
         entries.push(entry);
