+++ bathbot/src/commands/osu/osustats/globals.rs
@@ -527,7 +527,7 @@ async fn process_scores(
             rank,
             max_pp,
             stars: attrs.stars() as f32,
-            max_combo: attrs.max_combo() as u32,
+            max_combo: attrs.max_combo(),
         };
 
         entries.insert(i, entry);
