+++ bathbot/src/commands/osu/top/mod.rs
@@ -1272,7 +1272,7 @@ async fn process_scores(
             map,
             max_pp,
             stars: attrs.stars() as f32,
-            max_combo: attrs.max_combo() as u32,
+            max_combo: attrs.max_combo(),
         };
 
         if let Some(ref criteria) = filter_criteria {
