+++ bathbot/src/embeds/osu/player_snipe_stats.rs
@@ -69,7 +69,7 @@ impl PlayerSnipeStatsEmbed {
             let attrs = calc.performance().await;
             let stars = attrs.stars() as f32;
             let max_pp = attrs.pp() as f32;
-            let max_combo = attrs.max_combo() as u32;
+            let max_combo = attrs.max_combo();
 
             let pp = match oldest_score.pp {
                 Some(pp) => pp,
